+++ src/main/java/com/luan/getlib/dao/BookDao.java
@@ -22,7 +22,7 @@ public class BookDao {
             st.setString(2, book.getGenre());
             st.setInt(3, book.getAmount());
             st.setDouble(4, book.getValue());
-            st.setString(5, book.getParentalRating());
+            st.setInt(5, book.getParentalRating());
             
             return st.executeUpdate() == 1;
         } catch(SQLException e){
@@ -31,45 +31,29 @@ public class BookDao {
         }
     }
     
-    public static boolean updateBook(Book book){
-        try(Connection conn = Database.getConnection();
-            PreparedStatement st = conn.prepareStatement("UPDATE books SET amount = ?, value = ? WHERE id = ?")){
-            
-            st.setInt(1, book.getAmount());
-            st.setDouble(2, book.getValue());
-            st.setInt(3, book.getId());
-            
-            return st.executeUpdate() == 1;
-        } catch(SQLException e){
-            e.printStackTrace();
-            return false;
-        }
-    }   
-    
     public static Book findById(int id){
         try(Connection conn = Database.getConnection();
-            PreparedStatement st = conn.prepareStatement("SELECT * FROM books WHERE id = ?")){
+            PreparedStatement st = conn.prepareStatement("SELECT * FROM books WHERE book_id = ?")){
             
             st.setInt(1, id);
             
             ResultSet rSet = st.executeQuery();
-            if(rSet.next()) return new Book(rSet.getInt("id"),
+            if(rSet.next()) return new Book(rSet.getInt("book_id"),
                                             rSet.getString("title"),
                                             rSet.getString("genre"),
                                             rSet.getInt("amount"),
                                             rSet.getDouble("value"),
-                                            rSet.getString("parentalRating"));
+                                            rSet.getInt("parentalRating"));
             
         } catch(SQLException e){
             e.printStackTrace();
-            return null;
         }
         return null;
     }
     
     public static List<Book> findByTitle(String title){
         try(Connection conn = Database.getConnection();
-            PreparedStatement st = conn.prepareStatement("SELECT id, title FROM books WHERE title LIKE ? ORDER BY title ASC")){
+            PreparedStatement st = conn.prepareStatement("SELECT book_id, title FROM books WHERE title LIKE ? ORDER BY title ASC")){
             
             st.setString(1, "%" + title + "%");
             
@@ -78,7 +62,7 @@ public class BookDao {
             List<Book> books = new ArrayList<>();
             
             while(rSet.next()){
-                books.add(new Book(rSet.getInt("id"),
+                books.add(new Book(rSet.getInt("book_id"),
                                    rSet.getString("title")));
             }
             return books;
@@ -90,19 +74,34 @@ public class BookDao {
     
     public static List<Book> getAllBooks(){
         try(Connection conn = Database.getConnection();
-            PreparedStatement st = conn.prepareStatement("SELECT id, title FROM books ORDER BY title ASC")){
+            PreparedStatement st = conn.prepareStatement("SELECT book_id, title FROM books ORDER BY title ASC")){
             ResultSet rSet = st.executeQuery();
             
             List<Book> books = new ArrayList<>();
             
             while(rSet.next()){
-                books.add(new Book(rSet.getInt("id"),
+                books.add(new Book(rSet.getInt("book_id"),
                                    rSet.getString("title")));
             }
             return books;
         } catch(SQLException e){
             e.printStackTrace();
             return null;
+        }
+    }
+    
+    public static boolean updateBook(Book book){
+        try(Connection conn = Database.getConnection();
+            PreparedStatement st = conn.prepareStatement("UPDATE books SET amount = ?, value = ? WHERE book_id = ?")){
+            
+            st.setInt(1, book.getAmount());
+            st.setDouble(2, book.getValue());
+            st.setInt(3, book.getId());
+            
+            return st.executeUpdate() == 1;
+        } catch(SQLException e){
+            e.printStackTrace();
+            return false;
         }
     }
 }