+++ src/main/java/com/luan/getlib/interfaces/MenuScreen.java
@@ -1,13 +1,22 @@
 package com.luan.getlib.interfaces;
 
 import com.luan.getlib.dao.BookDao;
+import com.luan.getlib.dao.CustomerDAO;
 import com.luan.getlib.dao.EmployeeDAO;
+import com.luan.getlib.dao.OperationDAO;
+import com.luan.getlib.models.Address;
 import com.luan.getlib.models.Book;
 import com.luan.getlib.models.Customer;
 import com.luan.getlib.models.Employee;
+import com.luan.getlib.models.Operation;
 import com.luan.getlib.security.PasswordUtils;
+import com.luan.getlib.service.CurrencyService;
+import com.luan.getlib.service.RestCountriesService;
+import com.luan.getlib.service.ZipCodeService;
 import com.luan.getlib.utils.DataValidator;
 import com.luan.getlib.utils.InputReader;
+import java.time.LocalDate;
+import java.time.Period;
 import java.util.List;
 
 /**
@@ -16,17 +25,25 @@ import java.util.List;
  */
 public class MenuScreen {
     private static final InputReader sc = new InputReader();
-    private static String password, salt;
-    private static String title, genre, parentalRating;
-    private static int amount;
+    private static final ZipCodeService zip = new ZipCodeService();
+    private static RestCountriesService curr = new RestCountriesService();
+    private static CurrencyService currConvert = new CurrencyService();
+    private static String zipCode, nation, state, city, street, number, currency, email, phone, username, password, salt;
+    private static String title, genre;
+    private static int amount, parentalRating;
     private static double value;
     private static List<Book> books;
     private static Book book;
-    private static char exit = 'n';
+    private static char exit;
+    private static Address address;
+    private static Operation operation;
+    
+    
     public static void employee(Employee emp){
-        // TODO
+        
         System.out.println("Bem-Vindo " + emp.getFullName());
         while(true){
+            exit = 'n';
             System.out.println("Selecione uma opção:");
             System.out.println("1-Alterar senha");
             System.out.println("2-Listar livros");
@@ -37,12 +54,12 @@ public class MenuScreen {
                 case 0:
                     return;
                 case 1:
-                    System.out.print("Defina uma senha: ");
+                    System.out.print("Nova senha: ");
                     password = sc.getNextLine();
                     System.out.print("Confirme a senha: ");
                     while (DataValidator.isPasswordValid(password, sc.getNextLine())){
                         System.out.println("Tente novamente!");
-                        System.out.print("Defina uma senha: ");
+                        System.out.print("Nova senha: ");
                         password = sc.getNextLine();
                         System.out.print("Confirme a senha: ");
                     }
@@ -58,8 +75,8 @@ public class MenuScreen {
                     break;
                 case 2:
                     listBooks();
-                    System.out.println("Pressione [ENTER] para voltar!");
-                    while(sc.getNextLine().equals("\n"));
+                    System.out.println("Pressione [Qualquer tecla] para voltar!");
+                    sc.getNextLine();
                     break;
                 case 3:
                     // TODO
@@ -75,7 +92,7 @@ public class MenuScreen {
                                 System.out.print("Gênero: ");
                                 genre = sc.getNextLine();
                                 System.out.print("Classificação Indicativa: ");
-                                parentalRating = sc.getNextLine();
+                                parentalRating = sc.getNextInt();
                                 System.out.print("Quantidade: ");
                                 amount = sc.getNextInt();
                                 System.out.print("Valor: $");
@@ -114,7 +131,7 @@ public class MenuScreen {
                                 System.out.print("Quantidade a inserir: ");
                                 amount = sc.getNextInt();
                                 System.out.println("Nova quantidade: " + (book.getAmount() + amount));
-                                System.out.print("Confirma a operação (y/n)? ");
+                                System.out.print("Confirma a operação? (y/n)");
                                 switch(sc.getNext()){
                                     case 'y':
                                         book.setAmount(book.getAmount() + amount);
@@ -131,7 +148,7 @@ public class MenuScreen {
                                 System.out.print("Quantidade a remover: ");
                                 amount = sc.getNextInt();
                                 System.out.println("Livros restantes: " + (book.getAmount() - amount));
-                                System.out.print("Confirma a operação (y/n)? ");
+                                System.out.print("Confirma a operação? (y/n)");
                                 switch(sc.getNext()){
                                     case 'y':
                                         book.setAmount(book.getAmount() - amount);
@@ -147,7 +164,7 @@ public class MenuScreen {
                                 System.out.println("Valor anterior: $" + book.getValue());
                                 System.out.print("Novo Valor: $");
                                 value = sc.getNextDouble();
-                                System.out.println("Confirma a operação (y/n)?");
+                                System.out.println("Confirma a operação? (y/n)");
                                 switch(sc.getNext()){
                                     case 'y':
                                         book.setValue(value);
@@ -165,7 +182,213 @@ public class MenuScreen {
     }
     
     public static void customer(Customer cst){
-        // TODO
+        System.out.println("Bem-Vindo " + cst.getFullName());
+        while (true) {
+            exit = 'n';
+            System.out.println("Selecione uma opção: ");
+            System.out.println("1-Editar cadastro");
+            System.out.println("2-Listar livros");
+            System.out.println("3-Buscar livros");
+            System.out.println("0-Sair");
+            
+            switch(sc.getNextInt()){
+                case 0:
+                    return;
+                case 1:
+                    // TODO
+                    System.out.println("Selecione o que deseja alterar: ");
+                    System.out.println("1-Endereço");
+                    System.out.println("2-E-Mail");
+                    System.out.println("3-Telefone");
+                    System.out.println("4-Nome de usuário");
+                    System.out.println("5-Senha");
+                    System.out.println("0-Voltar");
+                    switch(sc.getNextInt()){
+                        case 0:
+                            break;
+                        case 1:
+                            if((address = zip.getAddressByZipCode(zipCode)) == null){
+                                System.out.println("Não foi possível encontrar o CEP fornecido! Preencha as informações manualmente!");
+                                System.out.print("País: ");
+                                nation = sc.getNextLine();
+                                System.out.print("Estado: ");
+                                state = sc.getNextLine();
+                                System.out.print("Cidade: ");
+                                city = sc.getNextLine();
+                                System.out.print("Rua: ");
+                                street = sc.getNextLine();
+                                System.out.print("Número: ");
+                                number = sc.getNextLine();
+                                address = new Address(nation, state, city, street, number, zipCode);
+                                currency = curr.getCurrencyByCountry(nation.replace(" ", "%20"), "name");
+                            } else {
+                                System.out.print("Rua: ");
+                                address.setStreet(sc.getNextLine());
+                                System.out.print("Número: ");
+                                address.setNumber(sc.getNextLine());
+                                currency = curr.getCurrencyByCountry(address.getNation(), "alpha");
+                            }
+                            
+                            cst.setAddress(address);
+                            cst.setCurrency(currency);
+                            break;
+                        case 2:
+                            System.out.print("Novo E-Mail: ");
+                            email = sc.getNextLine();
+                            cst.setEmail(email);
+                            break;
+                        case 3:
+                            System.out.print("Novo telefone: ");
+                            phone = sc.getNextLine();
+                            cst.setPhone(phone);
+                            break;
+                        case 4:
+                            System.out.print("Novo nome de usuário: ");
+                            while(!DataValidator.isUsernameValid(username = sc.getNextLine())){
+                                System.out.print("Novo nome de usuário: ");
+                            }
+                            cst.setUsername(username);
+                            break;
+                        case 5:
+                            System.out.print("Nova senha: ");
+                            password = sc.getNextLine();
+                            System.out.print("Confirme a senha: ");
+                            while(!DataValidator.isPasswordValid(password, sc.getNextLine())){
+                                System.out.println("Tente novamente!");
+                                System.out.print("Nova senha: ");
+                                password = sc.getNextLine();
+                                System.out.print("Confirme a senha: ");
+                            }
+                            
+                            salt = PasswordUtils.generateSalt();
+                            cst.setPassword(password = PasswordUtils.encryptPassword(password, salt));
+                            break;
+                    }
+                    updateCustomer(cst);
+                    break;
+                case 2:
+                    listBooks();
+                    System.out.println("Pressione [Qualquer tecla] para voltar!");
+                    sc.getNextLine();
+                    break;
+                case 3:
+                    System.out.print("Digite o nome do livro: ");
+                    title = sc.getNextLine();
+                    if(!listBooks(title)){
+                        System.out.println("A busca não encontrou nenhuma correspondência!");
+                        continue;
+                    } else {
+                        System.out.print("Digite o Id do livro desejado: ");
+                        if((book = BookDao.findById(sc.getNextInt())) != null){
+                            value = currConvert.convertFromUSD(book.getValue(), cst.getCurrency());
+                            double rentValue;
+                            do{
+                                System.out.println("\nId: " + book.getId());
+                                System.out.println("Título: " + book.getTitle());
+                                System.out.println("Gênero: " + book.getGenre());
+                                System.out.println("Classificação Indicativa: " + book.getParentalRating());
+                                System.out.println("Quantidade: " + book.getAmount());
+                                System.out.printf("Valor: %s %.2f\n\n", cst.getCurrency(), value);
+                                operation = OperationDAO.findRentByBookAndCustomer(book.getId(), cst.getId());
+                                
+                                if(operation != null){
+                                    rentValue = operation.calculateRentValue();
+                                    System.out.printf("1-Livro alugado em %s por %s %.2f / dia! Valor devido = %s %.2f! Devolver\n",
+                                                  operation.getOperationDate().toString(), cst.getCurrency(), operation.getValue(), cst.getCurrency(), rentValue);
+                                    System.out.println("2-Comprar");
+                                    System.out.println("0-Voltar");
+
+                                    switch(sc.getNextInt()){
+                                        case 0:
+                                            exit = 'y';
+                                            break;
+                                        case 1:
+                                            System.out.printf("Valor a pagar: %s %.2f\n", cst.getCurrency(), rentValue);
+                                            System.out.print("Confirma a devolução? (y/n) ");
+                                            switch(sc.getNext()){
+                                                case 'y':
+                                                    if(OperationDAO.deleteOperation(operation.getId())){
+                                                        System.out.println("Livro devolvido!");
+                                                        book.setAmount(book.getAmount() + 1);
+                                                        BookDao.updateBook(book);
+                                                    }
+                                                    break;
+                                                case 'n':
+                                                    break;
+                                            }
+                                            break;
+                                        case 2:
+                                            System.out.printf("Valor do livro: %s %.2f\n", cst.getCurrency(), value);
+                                            System.out.print("Confirma a compra? (y/n) ");
+                                            switch(sc.getNext()){
+                                                case 'y':
+                                                    operation.setType('p');
+                                                    operation.setOperationDate(LocalDate.now());
+                                                    operation.setValue(value);
+                                                    if(OperationDAO.updateOperation(operation)){
+                                                        System.out.println("Livro adquirido com sucesso!");
+                                                    }
+                                                    break;
+                                                case 'n':
+                                                    break;                                                    
+                                            }
+                                            break;
+                                    }
+                                } else {
+                                    if(Period.between(cst.getBirthDate(), LocalDate.now()).getYears() < book.getParentalRating()){
+                                        System.out.println("Você não possui idade suficiente para ler este livro!");
+                                        System.out.println("Pressione [Qualquer tecla] para voltar!");
+                                        sc.getNextLine();
+                                        break;
+                                    } else {
+                                        System.out.println("1-Aluguel");
+                                        System.out.println("2-Comprar");
+                                        System.out.println("0-Voltar");
+
+                                        switch(sc.getNextInt()){
+                                            case 0:
+                                                exit = 'y';
+                                                break;
+                                            case 1:
+                                                rentValue = value/100;
+                                                System.out.printf("O valor do aluguel deste livro é de %s %.2f / dia\n", cst.getCurrency(), rentValue);
+                                                System.out.print("Deseja confirmar o aluguel? (y/n) ");
+                                                switch(sc.getNext()){
+                                                    case 'y':
+                                                        if(OperationDAO.saveOperation(new Operation('r', book.getId(), cst.getId(), LocalDate.now(), rentValue))){
+                                                            System.out.println("Livro alugado com sucesso!");
+                                                            book.setAmount(book.getAmount() - 1);
+                                                            BookDao.updateBook(book);
+                                                        }
+                                                        break;
+                                                    case 'n':
+                                                        break;
+                                                }
+                                                break;
+                                            case 2:
+                                                System.out.printf("Valor do livro: %s %.2f\n", cst.getCurrency(), value);
+                                                System.out.print("Confirma a compra? (y/n) ");
+                                                switch(sc.getNext()){
+                                                    case 'y':
+                                                        if(OperationDAO.saveOperation(new Operation('p', book.getId(), cst.getId(), LocalDate.now(), value))){
+                                                            System.out.println("Livro adquirido com sucesso!");
+                                                            book.setAmount(book.getAmount() - 1);
+                                                            BookDao.updateBook(book);
+                                                        }
+                                                        break;
+                                                    case 'n':
+                                                        break;
+                                                }
+                                                break;
+                                        }
+                                    }
+                                }                                    
+                            } while (exit != 'y');
+                        }
+                    }
+                    break;
+            }
+        }        
     }
     
     private static void listBooks(){
@@ -196,5 +419,17 @@ public class MenuScreen {
         } else {
             System.out.println("Ocorreu um erro ao alterar o livro: " + book.getTitle() + "! Tente novamente mais tarde!");
         }
+    }
+    
+    private static void updateCustomer(Customer cst){
+        if(CustomerDAO.updateCustomer(cst)){
+            System.out.println("Operação concluída com sucesso!");
+        } else {
+            System.out.println("Occoreu um erro ao alterar o cadastro! Tente novamente mais tarde!");
+        }
+    }
+    
+    private static void purchaseBook(){
+        
     }
 }